- Implemented Refresh token API

+++ EzgIOS/Helpers/Global.swift
@@ -8,7 +8,7 @@ class Global {
     var baseUrl = "https://api.ezgolftech.com/api/v1/"
     var token: String = ""
     var refreshToken: String = ""
-    var tokenExpireAt: Int = 0
+    var tokenExpireAt: Double = 0
     var tokenType: String = ""
     var user: User?
